reading user input using two way bindings and handling form submission when clicking on calculate, we display the result array on the console
