+++ src/app/user-input/user-input.component.html
@@ -1,23 +1,23 @@
-<form>
+<form (ngSubmit)="calculate()">
     <div class = "input-group">
         <div class="input-label">
             <label>INITIAL INVESTMENT</label>
-            <input>
+            <input name="init_investment" [(ngModel)] = "initial_inv" />
         </div>    
         <div class="input-label">
-            <label>INITIAL INVESTMENT</label>
-            <input>
+            <label>ANNUAL INVESTMENT</label>
+            <input name="annual_investment" [(ngModel)] = "annual_inv"/>
         </div>  
     </div>
     <div class = "input-group">
         <div class="input-label">
-            <label>INITIAL INVESTMENT</label>
-            <input>
+            <label>EXPECTED RETURN</label>
+            <input name="expected_return" [(ngModel)] = "expected_ret"/>
         </div>  
         <div class="input-label">
-            <label>INITIAL INVESTMENT</label>
-            <input>
+            <label>DURATION</label>
+            <input name="duration" [(ngModel)] = "duration"/>
         </div>  
     </div>
-    <button>CALCULATE</button>
+    <button type="submit">CALCULATE</button>
 </form>
